+++ truffle/src/com.oracle.truffle.runtime/src/com/oracle/truffle/runtime/hotspot/HotSpotTruffleRuntime.java
@@ -40,9 +40,6 @@
  */
 package com.oracle.truffle.runtime.hotspot;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.invoke.MethodType;
 import java.lang.ref.Reference;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
@@ -50,7 +47,6 @@ import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
-import java.util.concurrent.ExecutionException;
 import java.util.function.Consumer;
 
 import com.oracle.truffle.api.CompilerDirectives;
@@ -175,9 +171,6 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
 
     private final HotSpotVMConfigAccess vmConfigAccess;
     private final int jvmciReservedLongOffset0;
-    private final int jvmciReservedReference0Offset;
-    private final MethodHandle setJVMCIReservedReference0;
-    private final MethodHandle getJVMCIReservedReference0;
 
     public HotSpotTruffleRuntime(TruffleCompilationSupport compilationSupport) {
         super(compilationSupport, Arrays.asList(HotSpotOptimizedCallTarget.class, InstalledCode.class, HotSpotThreadLocalHandshake.class, HotSpotTruffleRuntime.class));
@@ -189,37 +182,27 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
         try {
             longOffset = vmConfigAccess.getFieldOffset("JavaThread::_jvmci_reserved0", Integer.class, "jlong", -1);
         } catch (NoSuchMethodError error) {
-            // jvmci is too old to have this overload of getFieldOffset
-            longOffset = -1;
+            throw CompilerDirectives.shouldNotReachHere("This JDK does not have JavaThread::_jvmci_reserved0", error);
         } catch (JVMCIError error) {
             try {
                 // the type of the jvmci reserved field might still be old.
                 longOffset = vmConfigAccess.getFieldOffset("JavaThread::_jvmci_reserved0", Integer.class, "intptr_t*", -1);
             } catch (NoSuchMethodError e) {
-                longOffset = -1;
+                e.initCause(error);
+                throw CompilerDirectives.shouldNotReachHere("This JDK does not have JavaThread::_jvmci_reserved0", e);
             }
+        }
+        if (longOffset == -1) {
+            throw CompilerDirectives.shouldNotReachHere("This JDK does not have JavaThread::_jvmci_reserved0");
         }
         this.jvmciReservedLongOffset0 = longOffset;
-        this.jvmciReservedReference0Offset = vmConfigAccess.getFieldOffset("JavaThread::_jvmci_reserved_oop0", Integer.class, "oop", -1);
 
-        MethodHandle setReservedReference0 = null;
-        MethodHandle getReservedReference0 = null;
-        if (jvmciReservedReference0Offset != -1) {
-            installReservedOopMethods(null);
-
-            try {
-                setReservedReference0 = MethodHandles.lookup().findVirtual(HotSpotJVMCIRuntime.class,
-                                "setThreadLocalObject", MethodType.methodType(void.class, int.class, Object.class));
-                getReservedReference0 = MethodHandles.lookup().findVirtual(HotSpotJVMCIRuntime.class,
-                                "getThreadLocalObject", MethodType.methodType(Object.class, int.class));
-            } catch (NoSuchMethodException | IllegalAccessException e) {
-                /*
-                 * This is expected. Older JVMCI versions do not have setThreadLocalObject.
-                 */
-            }
+        int jvmciReservedReference0Offset = vmConfigAccess.getFieldOffset("JavaThread::_jvmci_reserved_oop0", Integer.class, "oop", -1);
+        if (jvmciReservedReference0Offset == -1) {
+            throw CompilerDirectives.shouldNotReachHere("This JDK does not have JavaThread::_jvmci_reserved_oop0");
         }
-        this.setJVMCIReservedReference0 = setReservedReference0;
-        this.getJVMCIReservedReference0 = getReservedReference0;
+
+        installReservedOopMethods(null);
     }
 
     public int getJVMCIReservedLongOffset0() {
@@ -298,17 +281,6 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
         }
     }
 
-    @Override
-    public boolean isLatestJVMCI() {
-        if (getJVMCIReservedReference0 == null) {
-            return false;
-        }
-        if (getJVMCIReservedLongOffset0() == -1) {
-            return false;
-        }
-        return true;
-    }
-
     /*
      * Used reflectively in CompilerInitializationTest.
      */
@@ -338,12 +310,12 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
                 this.initializeCallTarget = initCallTarget;
 
                 pendingTransferToInterpreterOffset = compiler.pendingTransferToInterpreterOffset(callTarget);
-                assert pendingTransferToInterpreterOffset != -1;
+                if (pendingTransferToInterpreterOffset == -1) {
+                    throw CompilerDirectives.shouldNotReachHere("Invalid offset for JavaThread::_pending_transfer_to_interpreter");
+                }
 
                 installCallBoundaryMethods(compiler);
-                if (jvmciReservedReference0Offset != -1) {
-                    installReservedOopMethods(compiler);
-                }
+                installReservedOopMethods(compiler);
 
                 truffleCompiler = compiler;
                 traceTransferToInterpreter = engine.traceTransferToInterpreter;
@@ -467,36 +439,13 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
         installCallBoundaryMethods((HotSpotTruffleCompiler) truffleCompiler);
     }
 
-    public MethodHandle getSetThreadLocalObject() {
-        return setJVMCIReservedReference0;
-    }
-
-    public MethodHandle getGetThreadLocalObject() {
-        return getJVMCIReservedReference0;
-    }
-
-    public boolean bypassedReservedOop(boolean waitForInit) {
-        if (jvmciReservedReference0Offset == -1) {
-            throw CompilerDirectives.shouldNotReachHere("bypassedReservedOop without field available. default fast thread locals should be used instead.");
-        }
-
+    public boolean bypassedReservedOop() {
         CompilationTask task = initializationTask;
         if (task != null) {
-            while (waitForInit) {
-                try {
-                    task.awaitCompletion();
-                    break;
-                } catch (ExecutionException e) {
-                    throw new AssertionError("Initialization failed.", e);
-                } catch (InterruptedException e) {
-                    continue;
-                }
-            }
             /*
-             * We were currently initializing. No need to reinstall the code stubs. Just try using
-             * them again has a very likely-hood of succeeding or if we do not wait for
-             * inititialization then the caller can use oop accessor methods
-             * (setJVMCIReservedReference0, getJVMCIReservedReference0) instead.
+             * We were currently initializing. No need to reinstall the code stubs. The caller can
+             * use oop accessor methods (setJVMCIReservedReference0, getJVMCIReservedReference0)
+             * instead.
              */
             return true;
         }
@@ -713,12 +662,7 @@ public final class HotSpotTruffleRuntime extends OptimizedTruffleRuntime {
 
     @Override
     protected AbstractFastThreadLocal getFastThreadLocalImpl() {
-        if (jvmciReservedReference0Offset != -1) {
-            return HotSpotFastThreadLocal.SINGLETON;
-        } else {
-            // fallback to default thread local
-            return null;
-        }
+        return HotSpotFastThreadLocal.SINGLETON;
     }
 
     public static HotSpotTruffleRuntime getRuntime() {